chore: Refactor test to make it usable by all queue implementations.
