--- lib/runtimes/priority-queue-test/processes/priorityQueue/configurationDefinition.ts
@@ -1,74 +1,0 @@
-import { Configuration } from './Configuration';
-import { ConfigurationDefinition } from '../../../shared/ConfigurationDefinition';
-import { getCorsSchema } from '../../../shared/schemas/getCorsSchema';
-import { getPortOrSocketSchema } from '../../../shared/schemas/getPortOrSocketSchema';
-import { getPriorityQueueStoreOptionsSchema } from '../../../shared/schemas/getPriorityQueueStoreOptionsSchema';
-import { getPublisherOptionsSchema } from '../../../shared/schemas/getPublisherOptionsSchema';
-import { getSubscriberOptionsSchema } from '../../../shared/schemas/getSubscriberOptionsSchema';
-
-const corsSchema = getCorsSchema(),
-      portOrSocketSchema = getPortOrSocketSchema(),
-      priorityQueueStoreOptionsSchema = getPriorityQueueStoreOptionsSchema(),
-      publisherOptionsSchema = getPublisherOptionsSchema(),
-      subscriberOptionsSchema = getSubscriberOptionsSchema();
-
-const configurationDefinition: ConfigurationDefinition<Configuration> = {
-  corsOrigin: {
-    environmentVariable: 'CORS_ORIGIN',
-    defaultValue: '*',
-    schema: corsSchema
-  },
-  healthCorsOrigin: {
-    environmentVariable: 'HEALTH_CORS_ORIGIN',
-    defaultValue: '*',
-    schema: corsSchema
-  },
-  healthPortOrSocket: {
-    environmentVariable: 'HEALTH_PORT_OR_SOCKET',
-    defaultValue: 3_001,
-    schema: portOrSocketSchema
-  },
-  missedItemRecoveryInterval: {
-    environmentVariable: 'MISSED_RECOVERY_INTERVAL',
-    defaultValue: 5_000,
-    schema: { type: 'number', minimum: 1 }
-  },
-  portOrSocket: {
-    environmentVariable: 'PORT_OR_SOCKET',
-    defaultValue: 3_000,
-    schema: portOrSocketSchema
-  },
-  priorityQueueStoreOptions: {
-    environmentVariable: 'PRIORITY_QUEUE_STORE_OPTIONS',
-    defaultValue: { type: 'InMemory', expirationTime: 30_000 },
-    schema: priorityQueueStoreOptionsSchema
-  },
-  pubSubOptions: {
-    environmentVariable: 'PUB_SUB_OPTIONS',
-    defaultValue: {
-      channelForNewItems: 'newItem',
-      subscriber: { type: 'InMemory' },
-      publisher: { type: 'InMemory' }
-    },
-    schema: {
-      type: 'object',
-      properties: {
-        channelForNewItems: { type: 'string', minLength: 1 },
-        subscriber: subscriberOptionsSchema,
-        publisher: publisherOptionsSchema
-      },
-      required: [ 'channelForNewItems', 'subscriber', 'publisher' ],
-      additionalProperties: false
-    }
-  },
-  crashHandlerTargetFile: {
-    environmentVariable: 'CRASH_HANDLER_TARGET_FILE',
-    defaultValue: '',
-    schema: {
-      type: 'string',
-      minLength: 1
-    }
-  }
-};
-
-export { configurationDefinition };
